Add support for generating Windows executables with Launch4j
This adds an Ant task which converts the compiled installer jar to a
native executable by wrapping the jar in that executable. This should
reduce user errors where Java is not associated correctly with the .jar
extension and it also displays a nice message if Java could not be
found.

I needed to change a few lines in the source since the code did not
compile with the options present in build.xml.