Avoid computing file hashes concurrently
Computing the hashes concurrently decreases overall hashing performance
since the OS has to do a lot of seek operations on the disk to allow all
hashing operations to operate in sequence. This is especially harmful on
HDDs.

This uses a standard Java lock to ensure that only onc hash is computed
at a time which improved hash performance immensely for me.